Refactor: move changes to custom dropdown component, keep original reusable component
