+++ workspaces/ballerina/ballerina-side-panel/src/components/editors/DropdownEditor.tsx
@@ -16,134 +16,35 @@
  * under the License.
  */
 
-import React, { useEffect, useState, useRef } from "react";
-import styled from "@emotion/styled";
+import React, { useEffect } from "react";
 
-import { Dropdown, Button, CheckBox, ThemeColors, TextField } from "@wso2/ui-toolkit";
+import { Dropdown } from "@wso2/ui-toolkit";
 
 import { FormField } from "../Form/types";
 import { capitalize, getValueForDropdown } from "./utils";
 import { useFormContext } from "../../context";
 import { SubPanel, SubPanelView } from "@wso2/ballerina-core";
-import { McpToolsSelection, McpTool } from "./McpToolsSelection";
 
 interface DropdownEditorProps {
     field: FormField;
     openSubPanel?: (subPanel: SubPanel) => void;
-    // Additional props for MCP tools functionality
-    serviceUrl?: string;
-    configs?: object;
-    rpcClient?: any;
-    onToolsChange?: (selectedTools: string[]) => void;
-    renderToolsSelection?: () => React.ReactNode;
-    newServerUrl?: string;
-    mcpTools?: { name: string; description?: string }[]; // <-- add this line
 }
 
-const DropdownStack = styled.div`
-    display: flex;
-    flex-direction: column;
-    width: 100%;
-`;
-
-const DropdownSpacer = styled.div`
-    height: 5px;
-`;
-
 export function DropdownEditor(props: DropdownEditorProps) {
-    const { field, openSubPanel, serviceUrl, configs, rpcClient, onToolsChange, newServerUrl } = props;
+    const { field, openSubPanel } = props;
     const { form } = useFormContext();
-    const { register, setValue, watch } = form;
-    const [mcpTools, setMcpTools] = useState<{ name: string; description?: string }[]>(props.mcpTools || []);
-
-    // Sync mcpTools state with props.mcpTools
-    useEffect(() => {
-        if (props.mcpTools) {
-            setMcpTools(props.mcpTools);
-        }
-    }, [props.mcpTools]);
-    const [selectedMcpTools, setSelectedMcpTools] = useState<Set<string>>(new Set());
-    const [loadingMcpTools, setLoadingMcpTools] = useState(false);
-    const [mcpToolsError, setMcpToolsError] = useState<string>("");
-    const toolSelection = watch(field.key);
-    const [localServiceUrl, setLocalServiceUrl] = useState<string>("");
+    const { register, setValue } = form;
 
     useEffect(() => {
-        if (newServerUrl && newServerUrl !== localServiceUrl) {
-            setLocalServiceUrl(newServerUrl);
-            console.log(">>> New server URL set:", newServerUrl);
+        // if field.key is "modelType" and value is not set and default value set. then setValue to default value initially
+        if (field.key === "modelType" && field.value === undefined && (field.defaultValue || field.placeholder)) {
+            setValue(field.key, field.defaultValue || field.placeholder);
         }
-    }, [newServerUrl]);
-
-    const handleToolSelectionChange = (toolName: string, isSelected: boolean) => {
-        const newSelectedTools = new Set(selectedMcpTools);
-        if (isSelected) {
-            newSelectedTools.add(toolName);
-        } else {
-            newSelectedTools.delete(toolName);
-        }
-        setSelectedMcpTools(newSelectedTools);
-        // Call the callback with the updated selection
-        props.onToolsChange?.(Array.from(newSelectedTools));
-    };
-
-    const handleSelectAllTools = () => {
-        let newSelectedTools: Set<string>;
-        if (selectedMcpTools.size === mcpTools.length) {
-            newSelectedTools = new Set();
-        } else {
-            newSelectedTools = new Set(mcpTools.map(tool => tool.name));
-        }
-        setSelectedMcpTools(newSelectedTools);
-        // Call the callback with the updated selection
-        props.onToolsChange?.(Array.from(newSelectedTools));
-    };
-
-    // Call onToolsChange whenever selectedMcpTools changes
-    useEffect(() => {
-        props.onToolsChange?.(Array.from(selectedMcpTools));
-    }, [selectedMcpTools]);
-
-    const showScopeControls = field.key === "toolsToInclude";
+    }, [field.defaultValue, field.placeholder]);
 
     // HACK: create values for Scope field
     if (field.key === "scope") {
         field.items = ["Global", "Local"];
-    }
-
-    if (showScopeControls) {
-        return (
-            <DropdownStack>
-                <Dropdown
-                    id={field.key}
-                    description={field.documentation}
-                    {...register(field.key, { required: !field.optional, value: getValueForDropdown(field) })}
-                    label={capitalize(field.label)}
-                    items={field.itemOptions ? field.itemOptions : field.items?.map((item) => ({ id: item, content: item, value: item }))}
-                    required={!field.optional}
-                    disabled={!field.editable}
-                    onChange={(e) => {
-                        setValue(field.key, e.target.value);
-                        field.onValueChange?.(e.target.value);
-                    }}
-                    sx={{ width: "100%" }}
-                    containerSx={{ width: "100%" }}
-                    addNewBtnClick={field.addNewButton ? () => openSubPanel({ view: SubPanelView.ADD_NEW_FORM }) : undefined}
-                />
-                <DropdownSpacer />
-                {toolSelection === "Selected" && (
-                    <McpToolsSelection
-                        tools={props.mcpTools ?? mcpTools}
-                        selectedTools={selectedMcpTools}
-                        loading={loadingMcpTools}
-                        error={mcpToolsError}
-                        onToolSelectionChange={handleToolSelectionChange}
-                        onSelectAll={handleSelectAllTools}
-                        serviceUrl={localServiceUrl}
-                    />
-                )}
-            </DropdownStack>
-        );
     }
 
     return (
