+++ workspaces/ballerina/ballerina-side-panel/src/components/editors/EditorFactory.tsx
@@ -41,6 +41,7 @@ import { ContextAwareRawExpressionEditor } from "./RawExpressionEditor";
 import { IdentifierField } from "./IdentifierField";
 import { PathEditor } from "./PathEditor";
 import { HeaderSetEditor } from "./HeaderSetEditor";
+import { CustomDropdownEditor } from "./CustomDropdownEditor";
 
 interface FormFieldEditorProps {
     field: FormField;
@@ -107,15 +108,19 @@ export const EditorFactory = (props: FormFieldEditorProps) => {
     } else if (field.type === "EXPRESSION" && field.key === "resourcePath") {
         // HACK: this should fixed with the LS API. this is used to avoid the expression editor for resource path field.
         return <TextEditor field={field} handleOnFieldFocus={handleOnFieldFocus} />;
+    } else if (field.type.toUpperCase() === "ENUM" && props.mcpTools) {
+        // TODO: this is a temporary solution to handle the enum field with MCP tools.
+        return <CustomDropdownEditor field={field} mcpTools={props.mcpTools} onToolsChange={props.onToolsChange} />;
     } else if (field.type.toUpperCase() === "ENUM") {
         // Enum is a dropdown field
-        return <DropdownEditor field={field} mcpTools={props.mcpTools} onToolsChange={props.onToolsChange} />;
+        return <DropdownEditor field={field} openSubPanel={openSubPanel} />;
     } else if (field.type === "FILE_SELECT" && field.editable) {
         return <FileSelect field={field} />;
+    } else if (field.type === "SINGLE_SELECT" && field.editable && props.mcpTools) {
+        // TODO: this is a temporary solution to handle the single select field with MCP tools.
+        return <CustomDropdownEditor field={field} openSubPanel={openSubPanel} newServerUrl={newServerUrl} mcpTools={props.mcpTools} onToolsChange={props.onToolsChange} />;
     } else if (field.type === "SINGLE_SELECT" && field.editable) {
-        // HACK:Single select field is treat as type editor for now
-        console.log(">>> Single select field is treated as type editor", field);
-        return <DropdownEditor field={field} openSubPanel={openSubPanel} newServerUrl={newServerUrl} mcpTools={props.mcpTools} onToolsChange={props.onToolsChange} />;
+        return <DropdownEditor field={field} openSubPanel={openSubPanel} />;
     } else if (!field.items && (field.key === "type" || field.type === "TYPE") && field.editable) {
         // Type field is a type editor
         return (
